#17 Refactor: 가로 휠 컴포넌트화로 대체
